refactor(agent): clean up pkg/agent directory structure
- Move Compactor interface and CompactionResult to pkg/context/compactor.go
- Move TruncateCompactHint to pkg/context/truncate_compact_hint.go
- Move agent:tool tag parsing functions to pkg/context/tool_tag.go
- Move toolLoopGuard and malformed tool call recovery to pkg/agent/tool_guard.go
- Delete unused priority.go and priority_test.go
- Use type alias in agent.go for backward compatibility
- Reduce loop.go from 1903 to 1708 lines
