+++ pkg/agent/tool_guard.go
@@ -1,0 +1,208 @@
+package agent
+
+import (
+	"context"
+	"fmt"
+	"strings"
+
+	agentctx "github.com/tiancaiamao/ai/pkg/context"
+	"github.com/tiancaiamao/ai/pkg/llm"
+	traceevent "github.com/tiancaiamao/ai/pkg/traceevent"
+	"log/slog"
+)
+
+// toolLoopGuard detects and prevents infinite tool call loops.
+type toolLoopGuard struct {
+	maxConsecutive int
+	maxPerToolName int
+
+	lastSignature  string
+	consecutiveRun int
+	toolCallTotals map[string]int
+}
+
+func newToolLoopGuard(config *LoopConfig) *toolLoopGuard {
+	if config == nil {
+		return nil
+	}
+	maxConsecutive := resolveLoopGuardLimit(config.MaxConsecutiveToolCalls, defaultLoopMaxConsecutiveToolCalls)
+	maxPerToolName := resolveLoopGuardLimit(config.MaxToolCallsPerName, defaultLoopMaxToolCallsPerName)
+	if maxConsecutive == 0 && maxPerToolName == 0 {
+		return nil
+	}
+	return &toolLoopGuard{
+		maxConsecutive: maxConsecutive,
+		maxPerToolName: maxPerToolName,
+		toolCallTotals: make(map[string]int),
+	}
+}
+
+func resolveLoopGuardLimit(value, defaultValue int) int {
+	if value < 0 {
+		return 0
+	}
+	if value == 0 {
+		return defaultValue
+	}
+	return value
+}
+
+// Observe checks tool calls for loop patterns.
+func (g *toolLoopGuard) Observe(toolCalls []agentctx.ToolCallContent) (bool, string) {
+	for _, tc := range toolCalls {
+		name := strings.ToLower(strings.TrimSpace(tc.Name))
+		if name == "" {
+			name = "unknown"
+		}
+		signature := name + ":" + hashAny(tc.Arguments)
+
+		if signature == g.lastSignature {
+			g.consecutiveRun++
+		} else {
+			g.lastSignature = signature
+			g.consecutiveRun = 1
+		}
+
+		if g.maxConsecutive > 0 && g.consecutiveRun > g.maxConsecutive {
+			return true, fmt.Sprintf("detected %d consecutive identical tool calls (%s)", g.consecutiveRun, name)
+		}
+
+		g.toolCallTotals[name]++
+		if g.maxPerToolName > 0 && g.toolCallTotals[name] > g.maxPerToolName {
+			return true, fmt.Sprintf("tool %q called %d times in one run", name, g.toolCallTotals[name])
+		}
+	}
+	return false, ""
+}
+
+// ResetToolCount resets the call counter for a specific tool.
+// This is used when a tool call is productive (e.g., writing llm context)
+// and should not contribute to loop detection.
+func (g *toolLoopGuard) ResetToolCount(toolName string) {
+	if g == nil {
+		return
+	}
+	g.toolCallTotals[toolName] = 0
+}
+
+func sanitizeMessageForToolLoopGuard(msg *agentctx.AgentMessage, reason string) {
+	if msg == nil {
+		return
+	}
+
+	filtered := make([]agentctx.ContentBlock, 0, len(msg.Content)+1)
+	for _, block := range msg.Content {
+		switch block.(type) {
+		case agentctx.ToolCallContent:
+		default:
+			filtered = append(filtered, block)
+		}
+	}
+	filtered = append(filtered, agentctx.TextContent{
+		Type: "text",
+		Text: "\n\n[Loop guard] Stopped repeated tool execution to prevent an infinite loop.\nReason: " + strings.TrimSpace(reason),
+	})
+	msg.Content = filtered
+	msg.StopReason = "aborted"
+}
+
+func maybeRecoverMalformedToolCall(
+	ctx context.Context,
+	agentCtx *agentctx.AgentContext,
+	newMessages *[]agentctx.AgentMessage,
+	stream *llm.EventStream[AgentEvent, []agentctx.AgentMessage],
+	msg *agentctx.AgentMessage,
+	recoveryCount *int,
+) bool {
+	if msg == nil || agentCtx == nil || recoveryCount == nil {
+		return false
+	}
+	shouldRecover, reason := shouldRecoverMalformedToolCall(msg)
+	if !shouldRecover {
+		return false
+	}
+	if *recoveryCount >= defaultMalformedToolCallRecoveries {
+		slog.Warn("[Loop] malformed tool-call recovery limit reached",
+			"recoveryCount", *recoveryCount,
+			"reason", reason)
+		return false
+	}
+
+	*recoveryCount = *recoveryCount + 1
+	recoveryMsg := buildMalformedToolCallRecoveryMessage(reason, *recoveryCount)
+	agentCtx.Messages = append(agentCtx.Messages, recoveryMsg)
+	if newMessages != nil {
+		*newMessages = append(*newMessages, recoveryMsg)
+	}
+	if stream != nil {
+		stream.Push(NewToolCallRecoveryEvent(ToolCallRecoveryInfo{
+			Reason:  reason,
+			Attempt: *recoveryCount,
+		}))
+	}
+	traceevent.Log(ctx, traceevent.CategoryTool, "malformed_tool_call_recovery",
+		traceevent.Field{Key: "attempt", Value: *recoveryCount},
+		traceevent.Field{Key: "reason", Value: reason},
+	)
+	slog.Warn("[Loop] malformed tool call recovered",
+		"attempt", *recoveryCount,
+		"reason", reason)
+	return true
+}
+
+func shouldRecoverMalformedToolCall(msg *agentctx.AgentMessage) (bool, string) {
+	if msg == nil || len(msg.ExtractToolCalls()) > 0 {
+		return false, ""
+	}
+
+	if msg.StopReason == "tool_calls" {
+		return true, "stop_reason=tool_calls but no parsable tool call was produced"
+	}
+
+	text := strings.TrimSpace(msg.ExtractText())
+	thinking := strings.TrimSpace(msg.ExtractThinking())
+
+	candidates := []struct {
+		source string
+		text   string
+	}{
+		{source: "text", text: text},
+		{source: "thinking", text: thinking},
+	}
+
+	for _, candidate := range candidates {
+		body := strings.TrimSpace(candidate.text)
+		if body == "" {
+			continue
+		}
+
+		issues := DetectIncompleteToolCalls(body)
+		if len(issues) > 0 {
+			return true, fmt.Sprintf("%s: %s", candidate.source, strings.Join(issues, "; "))
+		}
+
+		lower := strings.ToLower(body)
+		if strings.Contains(lower, "<tool_call") ||
+			strings.Contains(lower, "<tool>") ||
+			strings.Contains(lower, "ErrorException") ||
+			strings.Contains(lower, " excer ") {
+			return true, fmt.Sprintf("%s: detected tool-call markup without a valid parsed tool call", candidate.source)
+		}
+	}
+
+	return false, ""
+}
+
+func buildMalformedToolCallRecoveryMessage(reason string, attempt int) agentctx.AgentMessage {
+	cleanReason := strings.TrimSpace(reason)
+	if cleanReason == "" {
+		cleanReason = "unknown parse failure"
+	}
+
+	text := fmt.Sprintf(
+		"[agentctx.Tool-call recovery, attempt %d] Your previous response attempted a tool invocation but the tool call format was invalid (%s). Re-emit the intended call using valid tool/function-call syntax only. If no tool is needed, provide the final answer directly.",
+		attempt,
+		truncateLine(cleanReason, 220),
+	)
+	return agentctx.NewUserMessage(text).WithVisibility(true, false).WithKind("tool_call_repair")
+}
